+++ azure/src/main/scala/org/broadinstitute/dsde/workbench/azure/ServiceBusMessageUtils.scala
@@ -1,0 +1,13 @@
+package org.broadinstitute.dsde.workbench.azure
+
+import com.azure.messaging.servicebus.ServiceBusReceivedMessage
+import com.google.protobuf.Timestamp
+import org.broadinstitute.dsde.workbench.model.TraceId
+
+import java.time.Instant
+
+object ServiceBusMessageUtils {
+
+  def getTraceIdFromCorrelationId(message: ServiceBusReceivedMessage): Option[TraceId] =
+    Option(message.getCorrelationId).map(new TraceId(_))
+}