+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GooglePublisher.scala
@@ -6,6 +6,7 @@ import com.google.pubsub.v1.PubsubMessage
 import fs2.Pipe
 import io.circe.Encoder
 import org.broadinstitute.dsde.workbench.model.TraceId
+import org.broadinstitute.dsde.workbench.util2.messaging.CloudPublisher
 import org.typelevel.log4cats.StructuredLogger
 
 trait GooglePublisher[F[_]] {
@@ -39,9 +40,17 @@ trait GooglePublisher[F[_]] {
 }
 
 object GooglePublisher {
+  @deprecated("Use cloudPublisherResource instead", "0.36")
   def resource[F[_]: Async: StructuredLogger](
     config: PublisherConfig
   ): Resource[F, GooglePublisher[F]] =
+    for {
+      publisher <- GooglePublisherInterpreter.publisher(config)
+    } yield GooglePublisherInterpreter(publisher)
+
+  def cloudPublisherResource[F[_]: Async: StructuredLogger](
+    config: PublisherConfig
+  ): Resource[F, CloudPublisher[F]] =
     for {
       publisher <- GooglePublisherInterpreter.publisher(config)
     } yield GooglePublisherInterpreter(publisher)